+++ CHANGELOG.md
@@ -1,2 +1,21 @@
 ### v1.0.0
 * Initial release
+
+### v1.0.1
+* Fixed Enable/Disable All buttons not saving to file.
+
+### v1.0.2
+* Purple items should not break things anymore.
+
+### v1.0.3
+* Made optimizations (to a few of my mods) to reduce load.
+* Updated Installation Instructions.
+
+### v1.0.4
+* Fixed being able to toggle in-run on the pause menu by restricting toggling to the character select screen.
+
+### v1.0.5
+* Updated all my mods to use a single helper file.
+* Updated installation instructions to use the mod manager instead.
+* Rewrote around new item functions in helper.
+    * Should be way less laggy post-loop.